Add Binary and Assignable traits
Binary is trait that enables abstraction over a binary operation.
Assignable is a helper trait for Binary that allows it to assign the
result of the binary operation the the left-hand operand.

This change also adds implementations of Add and Sub Binary operations
and accompanying add and sub functions that take two Futures and add and
sub their results respectively.